Only write to the db sometimes in Petition#increment_signature_count!
Even after moving writing signature_count to memcached the db experiences many locks on the petition rows because we timestamp the petition after each signature is validated.  While writing a value is less effort than incrementing an existing value, it's still producing row locks at high load.

We still need to write some timestamps, for example when the signature means the petition goes over some thresholds or would change state so we can't remove the writes entirely.  We also need to periodically timestamp the petition if it's being signed to flush the caches on the frontend which are based on the updated_at timestamp.  We settle on always updating "young" petitions on every signature, but ease off on "mature" petitions based on when the petition was last signed.

This does mean stamps will be out for a while, and there's an edgecase where the last two people to sign the petition do so within 10 seconds of each other so we cache the count at 1 less than it should be.  We decided we can live with that for now and if this becomes a permanent solution we can have periodic jobs that make sure to sync the timestamps with reality.

Note we have no specs for this, but have left unimplemented stubs to remind us if we make this more permanent.