Stop updating popular petitions every 10s
Even only updating a popular petitions timestamps every 10 seconds can produce locking on the row under heavy load, so for now we just disable the update entirely for these petitions.  Ideally we'd be able to ease off based on the signature rate, but this could be a hard query to run and we need signing to be fast.

To get around the issue of our petitions not actually getting updated timestamps we are running a terminal session to touch the petitions we know are under heavy traffic every 10seconds:

   while true
     puts "Running at #{Time.now}"
     Petition.where(id: [<list of petition ids>]).each { |p| p.touch }
     sleep 10
   end

This is obviously far from ideal, but it's something we can live with for now.