Look at updated_at on signatures, not petitions in cache reset
Now that one of the figures cached instead of written to the db is the petition's updated_at it's no longer correct to use that to find petitions that have been updated and might need their cached values reset.  Instead we look for petitions that have any signatures that have been updated in the window and use that instead.  We only look for validated signatures because, for now at least, the caching only happens when a signature is validated.