Avoid setting state twice in increment_signature_count!
When a petition is collecting sponsors but the creator has not yet validated their own email and we get to the sponsorship threshold we were trying to set the state to both 'validated' and 'sponsored'.  Postgres will complain if you try to set the same column multiple times in a single update and so the update would fail.  Instead of trying to set it twice we make sure to only set it once.

Note that it's not a huge problem if a creator doesn't directly validate their signature.  We know that their email is real if sponsors start coming in as the link for sponsors only appear in the email we send the creator.  If people are hitting that url then we know the email has been delivered so it's ok to go from 'pending' straight to 'sponsored'.