Stop loading the petition to check if a signature is a sponsor
Checking via the petition and it's links to the sponsors table might be more "correct" but it means we load the petition when we don't need to.  We instead use the association on the signature to run an exists? query - this also avoids trying to load the sponsor object when we might not need it.