Move updating counts after signing to a background process
Under heavy load we get contention on trying to write to the petitions table to update the signature_count and on trying to write to the 'GB' country journal.  We can control the contention by moving these updates to a background job and then scaling the number of workers.

We also move the email notifications that are sent to the creator about sponsors into the same job, as whether or not to send these emails relies on the data being updated.

Note that doing this requires telling some specs that use `Signature#validate!` to run background processes.  I only updated the ones that failed, rather than putting in place a solution for all specs - particularly as the background job might also send an email.