Add atomic dalli store to prevent thundering herds
When the site is under heavy load and the home page fragment expires we
can get a 'thundering herd' effect on the trending petitions query which
is slow and causes more contention.

To eliminate this problem the commit adds a custom cache store that writes
two keys - the original which adjusted to expire a little later and a
second TTL key which has the original expiry time. When this TTL key
expires the store optimistically locks the TTL key and if it fails it
return the stale content whereas if it succeeds it returns nil, triggering
the yield to the block in `fetch`.