Handle serialization errors in petition signed data update job
We don't queue these jobs during a creation event, so if a signature can't be deserialized it means that the signature has actually been deleted by an admin. This can't be recovered from so there's no reason to throw the exception and have the job retry several times.  We catch the exception, notify appsignal and let it complete gracefully.