Be more strict about content negotiation
We see a handful of errors for ActionView::MissingTemplate where people are requesting pages as json or xml when we don't support that.  By using respond_to and respond_with across all our public facing actions that render content we make sure that this sort of thing will be handled with a 406 response instead of a 500 error.

We introduce tests for all the urls that will render something, including failed posts to some create actions.  These tests are fairly slow, but I suspect that's just the cost of doing these sorts of full-end-to-end tests.  These tests also uncovered a couple of issues that we fix:

* browserconfig.xml and manifest.json could be requested without an extension
* feedback had an index action when it really needed a new action
* we needed to add explicit controller actions in places to force content negotiation rather than relying on falling through to the view template lookup