Increase the time the script will wait to enter standby
A number of deploys recently have failed to deregister from the ELB
properly so the ApplicationStop phase fails but the deploy carrying
on leading to the app serving requests from stale code.

The ELB connection draining timeout was reduced but didn't resolve
the problem so seeing whether increasing the timeout in the script
instead to see if that eliminates the problem.