Migrate email sent receipts to signatures
Doing a join between two tables with tens of millions of rows means that
iterating over the signatures that need emailing can take considerable
time for a large petition (e.g. 16 hours for 4.1m signatures). If we move
the timestamp columns to the signatures table there will be an initial
slow query but each successive batch of 1000 records for find_each will
be much quicker (approx. 3ms).

Before deploying we need to ensure that all existing bulk email jobs have
completed. After deployment we then need to run the migration task to
copy the timestamps from email_sent_receipts table - the table itself
will be deleted in a later deploy.