Move admin flash messages to locale file
To streamline the translation of the I18n keys override redirect_to so
that symbol keys are translated. If substitution is required then pass
an array with the last element being the substitution hash.

Also override render so that we can use :alert and :notice in a similar
fashion to how redirect_to uses them with the exception that they go
into flash.now and not the standard flash.

As part of this some logic around admin users was pushed down into the
model instead of the controller so that we can use model based I18n.