Expose fraudulent domains on petition admin page
List the domains marked as fraudulent on the petition admin page so
that the moderation team can see whether there's been significant
attempts at fraud on the petition.

The functional index for the domain ensures that the GROUP BY and
COUNT are performant even when there's a large number of signatures.

Previously we've eschewed an index on state since it didn't provide
much specificity but that has changed with the addition of the new
invalidated and fraudulent states.