Don't count invalidated signatures in trending query
Signatures that have been invalidated don't clear the validated_at
column for reasons of maintaining useful information in tracking
down any problems that may occur. However this means that the
trending query must exclude them so that it is accurate.

The reason that the check is `invalidated_at IS NULL` rather than
the more obvious `state = 'validated'` is for reasons of performance
- there are a lot of validated rows so the index isn't much help
in selecting rows whereas the invalidated_at check is a filter on
the already reduced rows of the validated_at index scan.