Fix edge cases when closing petitions
When trying to calculate the opening date for petitions that should be
closed we need to take account of the fact that date calculations using
months is inherently imprecise. This only really has an effect when the
date is the 31st though for August it's a much bigger effect since it
could affect 29th, 30th and 31st depending on whether it's a leap year.

The fix itself is straightforward - just add a day to take us onto the
1st of the next month. This does mean that we're running a background
job pointlessly for 1-2 days in August depending on whether it's a leap
year but the extra code to handle this condition isn't worth it.