Fix race condition in petition rejection
When a petition is rejected by two moderators at the same time there
is a race condition which results in an ActiveRecord::RecordNotUnique
error for the `has_one :rejection` association.

To fix this we rescue the error, force the association to reload which
picks up the existing record and then update that instead of inserting
our own. The result of this is that the last committer wins which is
the standard pattern elsewhere in the app with write conflicts.

The test for the race condition fakes two moderators by pre-loading the
rejection association which convinces Active Record that there is no
existing record so will not attempt to delete it first.