Fix race condition in government response controller
When two moderators update the government response simultaneously there
is a race condition which results in an ActiveRecord::RecordNotUnique
error for the `has_one :government_response` association.

To fix this we rescue the error, force the association to reload which
picks up the existing record and then update that instead of inserting
a new one. The result of this is that the last committer wins which is
the standard pattern elsewhere in the app with write conflicts.