Output UTF-8 encoded original url for Open Graph url tag
Sometimes the url returned by Rack is encoded as binary due to clients
not properly encoding urls. When this url was being output as a url tag
in the Open Graph meta tags it was causing the transformation of the
output buffer from UTF-8 to binary and causing later concatenations to
blow up with an incompatible character encodings exception.