Increase allowed size of government responses
The limit of 5000 doesn't take into account the fact that posting the
form sends CR/LFs which counts as two characters but in Word it's seen
as one character by the counter in the document chrome.

Fix this by increasing the validation limit to 6000 characters but
still instruct staff to keep responses to 5000 characters.