Add gathering support page for a petition
Some people were navigating to a petition page before it had gathered
enough sponsors for moderation and getting confused because we did not
discern between the two states and presented the same page for both
that said 5 people had supported the petition. To clarify this this
commit adds a 'Gathering Support' page for the situation where a
petition has yet to receive 5 signatures from sponsors.

https://www.pivotaltracker.com/story/show/120107273