Add an all local petitions view
We get a lot of requests for top petitions in a constituency but the
current page only shows open petitions. To reduce the need for running
custom database queries we can add a page to show all petitions and
then direct any FoI requests to the page for the constituency.