Add fragment caching to local petitions page
The queries for the local results pages can be quite heavy so cache
the generated HTML for 5 minutes to reduce the possibility of a DDoS
attack. As part of this the Petition model methods were changed to
return scopes instead of arrays otherwise there would be little benefit
from caching as the queries would've already been executed in the
controller action.