Update to use new Registers API response format
The response you receive from a request for the `/records` endpoint is
changing. This will be a breaking change. This change provides support
for both the old and new response style whilst we transition.

The biggest change here is the addition of a list of items under the
`item` [sic] key within each record. For the Country Register this will
only contain a single item.