Fix display of closing date
The `Petition#close!` method was taking the time that the job ran
as the closing time of the petition. Whilst this is technically
accurate it's confusing when displayed the next day because people
see the petition as closed when it's still that day.

Fix this by using the deadline when closing the petition.