Remove duplicated signature warning
When attempting to sign a petition, there is a validation step ran in
the form submission to prevent people from signing the petition multiple
times using the same email. This appears in the form of an error
saying "You cannot sign this petition again". This presents a privacy
concern to users and leaks information about which emails (and email
owners) have signed what petitions, as highlighted in
https://andrewhickey.info/2017/04/17/security-issue-with-parliament-petitions-website/.

This commit mitigates this by delaying the validation step into the
controller checks when creating a signature. When attempting to sign
a petition a second time (using the same details), the owner of the
email will receive an email alerting them their email was used in a
duplicated signature attempt. To prevent spamming inboxes, an
`email_count` attribute has been added to the signature table to keep a
track of emails sent to the owner of the email. The duplicate signature
email as well as the confirmation email sent out for the first time will
both increase this counter, and neither will be sent out when the
counter hits 5.