First stage of closing the website for the GE 2017
This adds dissolved messaging for when the dissolution has occurred
and prevents the creation of new petitions after that date. Since
archiving of the petitions and signature statistics can't happen
instantaneously we'll continue to show the petitions and popular
local petitions until that process is complete. Also adds a link on
the home page to the Register to Vote service.