Add signature email uuid
We want to be able to count the unique number of signature email
addresses but still anonymise them so add a uuid column that contains
a RFC 4122 v5 UUID using the URL namespace. The column defaults to
nil so that the operation can be performed without taking the site
down and then we run a job to backfill the existing signatures.