Only validate emails on create
Previously we had allowed '+' style addressing but it was removed
due to fraud concerns. However the validation for this now causes
a petition save to fail if the original email address is now valid
and the creator_signature association is loaded.

We can fix this by moving the email validations to on creation only
since the email address shouldn't be updated other than when it is
anonymised after the 12 month window.