Add clarification about early closing of petitions
The default closed message confusingly said that all petitions ran for
six months which obviously wasn't the case for the petitions that
were closed early due to the general election. Since we know that
petitions had their closed_at set to the exact time of the dissolution
we can compare that and eliminate other possible reasons for closing
a petition early.

This currently uses the Parliament singleton but we'll need to switch
to using a belongs_to association when we archive the petition.