Refactor signature uuid backfill
It's big ask for Active Record to iterate over 38 million rows
and update them within the maximum runtime for a delayed job so
refactor to run the task as a sequence of small jobs. This allows
us to shutdown workers and deploy a new version and the task will
pick up where it left off.