Use a curated list for example postcodes
There are too many disagreements between MapIt and the Parliament API
about whether a postcode falls within a constituency or not so use
a curated list and remove the dependency on a MapIt server.

The constituency reference data feed has every constituency that has
ever existed so we need to filter it to the current 650 that have no
end date in the response.