Fix clearing the petition notes field
We were validating the presence of the details field but this this
meant that a petition couldn't be untagged if the notes just consisted
of tags. To work around this moderators would fill the notes field
with a single '.' but we should just not validate the presence of the
details attribute.