Reset column information in migration
When running `rake db:migrate` from scratch Rails will load the column
information for a table and cache it so when the model is used a second
time the column information is stale and a previously added column is
not found.

Fix this by resetting the column information before using the model(s).