Store journal data in JSON columns
For current petitions we store individual journal records for each
country and constituency for performance reasons. For archived
petitions there is no need to do this as the data is static. However
the MP details can still change so we need to cache just the basic
information and then recreate the full array when fetching it from
the database.