Fix test setup for closing petitions
The Site.open_at_for_closing method gives a timestamp at which the
open_at column needs to be less than to be considered for closing.
When the date is the end of the month a setup of 2 days ago fails,
especially in August because there are periods longer than that for
which no petition will be closed. Change the setup to subtract a
day instead which should always give us an open_at which is due for
closure whatever day of the year the tests are run.