Lock signature when archiving
Although nominally with a single worker processing a single job it
shouldn't happen, if the job runs over the max runtime it can start
running the new job whilst still running the old one. This leads
to a race condition and record not unique errors.

To protect against this we can lock the signature record to ensure
that another process running the same job can't archive the signature
at the same time and also only process 1000 signatures at a time
before scheduling a new job which gives us protection against hitting
the max runtime limit of Delayed::Job.