Convert circular relationship into a has_one
Cycles in object graphs are best avoided so convert the creator
signature association to a has_one association using a boolean field
to indicate that it's a creator.

To prevent duplicate creator signatures we add a partial unique index
on the signature petition_id column with a `creator = 't'` condition.