Prevent boolean argument to prepared statements
When using hash conditions to `where`, Active Record will treat that
as an argument for the prepared statement even though it will never
change in reality. This causes problems for our partial indexes on
`creator` and `sponsor` columns in the signatures and the
archived_signatures tables because of the following behaviour as
documented in PREPARE[1]:

    Prepared statements can use generic plans rather than
    re-planning with each set of supplied EXECUTE values.
    This occurs immediately for prepared statements with no
    parameters; otherwise it occurs only after five or more
    executions produce plans whose estimated cost average
    (including planning overhead) is more expensive than the
    generic plan cost estimate. Once a generic plan is chosen,
    it is used for the remaining lifetime of the prepared statement.

What this means in effect is that after each Active Record connection
was used five times to fetch a petition creator it was defaulting to
the generic query plan which in our case turned out to be a sequential
scan of the signatures table.

By using an opaque Arel equality node passed to `where` we can
prevent Active Record from seeing that as an argument to a prepared
statement so the plan used by PostgreSQL continues to use our partial
index even after the statement has been used five times.

[1]: https://www.postgresql.org/docs/9.6/static/sql-prepare.html#SQL-PREPARE-NOTES