Harden bulk signature actions against attack
Use a per-session secret key to encrypt and sign a list of ids presented
to the moderator and ensure that when the bulk action is requested the
selected set of ids is contained within the signed list of ids and that
the list of ids hasn't been tampered with.

This provides an additional layer of security should any vulnerabilities
be discovered within the authlogic gem.

Also limit the number of selected ids accepted by the bulk action to a
maximum of fifty. However, this is somewhat redundant given the fact
that there will be a maximum of fifty ids in the signed list of ids.