Enforce UTF-8 encoding for token parameters
Although the encoding was technically already being forced by Rails
the param could still have a invalid character which blew up when
passed to PostgreSQL as a `PG::CharacterNotInRepertoire` exception.

Fix this by replacing invalid characters with a `�` and also don't
pass the token param to the Active Record finder method, just compare
it in memory instead.