Use Time.current for travelling with a duration
The Rails test helper `travel` adds the duration to `Time.now` and
not `Time.current` which means that for about a week this year there's
a difference between adding 1 year to Travis CI's default UTC system
time vs. adding 1 year to the application's time in the Europe/London
timezone because the clocks going forward an hour is a week earlier.