Make signature verification idempotent
Sometimes people are confused when they click the link in the email
and they get the message that they've already signed the petition
because their email scanner has already followed the link and verified
the signature.

There are two ways we can address this problem:

1) Add a second step to the verification process,
   e.g. a form on the page that submits via POST

2) Make repeated clicks idempotent,
   e.g. the page presented to the user is the same every time

We choose the latter because the intent of the email is verification
of the address and not to give the user another chance to avoid signing
as there are multiple steps to the process anyway. Given this, the fact
that the email scanner verified the address is unimportant. Also
introducing a second step would be a significant change in UX and would
possibly confuse people that have signed petitions previously.

It's often stated that having a side-effect by clicking a link is
breaking REST but GET actions (along with HEAD, PUT and DELETE) are
defined as idempotent in [RFC 7231][1] and not nullipotent, therefore
it's perfectly acceptable to validate the signature in response.

[1]: https://tools.ietf.org/html/rfc7231#section-4.2.2