Reduce confusion when people share their signed page
Although we embed sharing links in the page people often share the last
page in the signing process (/signatures/:id/signed?token=123456) which
thanks them for signing a petition. When someone else clicks this link
they also see the same page thanking them for signing the petition which
has led to some confusion and emails from people saying they just wanted
to look at the petition before signing it.

To fix this we're adding a hidden token which gets set in the session
when the person follows the verification link and removing the token
from the url. This means that when someone shares that url and someone
else clicks it they don't have the hidden token in the session and we
can redirect to the petition page or then moderation info page if the
petition isn't published yet.