Add statistics about duplicate email usage
Because people sometimes share email addresses we allow up to two
signatures using the same email address as long as the name is different
and the postcode is the same. The petitions committee often receives
queries about the abuse of this rule so to improve the speed of handling
these add the statistics about them to the petition admin page.

Because the query takes a long time with petitions that have a large
number of signatures the implementation is a little complicated and
we have to cache the results and update them every hour. If a more
recent number is required the sysadmin user can kick off an manual
refresh of the statistics. The pattern is flexible so that we can add
further statistics in the future.