Move signature count updating to a single task
Lots of +1 queries are preventing the app from reading the row of
a popular petition so batch these updates into one query by taking
advantage of the validated_at timestamp on signatures to find
the signatures validated since the previous update.

Also provides the ability to control the rate of updates via the
admin backend and to disable it altogether.