Use a historical time window for updating signature counts
Since the timestamps are coming from ruby they're subject to drift
meaning that some signatures will be inserted with a validated_at
timestamp in an out of order fashion. By setting the window to look
at the previous window hopefully any such signatures will have all
been committed by then and if they haven't it's likely the database
is in flames anyway and we'll need to re-sync the counts.