Allow resetting a signature count with counting enabled
Having to stop the signature counts being updated whilst a petition
is having its signature count reset is big limitation when a problem
occurs that requires many petitions to be reset. By allowing the
signature count update job to skip those petitions being reset we
can remove the limitation and not have to rush to complete them.