Only update journals if the main signature count is updated
There's a subtle race condition when a petition is created and the
first sponsor visits the new signature form. This action validates
the creator signature but because the batch counting job runs a few
seconds behind it means that the creator signature can be seen by
the journal increment methods until that first signature is counted
and last_signed_at set on the petition.