Allow overriding of the default sprockets cache path
Sprockets 3.0 and later uses Base64 digests for cache paths and these
present problems when used with NTFS on Windows which is case-insensitive.

By allowing the overriding of the cache path from an environment variable
we can use a Docker volume to cache the asset compilation. This is actually
faster on both Mac and Windows than the default path.