Translate routes into Welsh
This implements a simple mapper that supports `scope`, `get` and `post`
so that we can capture the routes in English and then translate the
paths and suffix the route names. It then adds custom url helpers to
select the correct url helper based on the current locale. It also then
adds custom url redirects to handle English paths sent to the Welsh
domain and vice-versa.