Autocorrect a select number of high-traffic domains
Initially looked at using a Levenshtein algorithm to calculate edit
distance and autocorrect based on that but there are too many false
positives even when capping the edit distance to a maximum of 1.

Instead, we analysed a corpus of emails to find the most common errors
for our highest-traffic domains and then generated regular expressions
to correct those errors. We only autocorrect on the first submission
of the email which presents the user a chance to correct any false
positives when we replay their email back to them.