Add the ability to pause signature collection
There may be scenarios in which the Petitions service will have to
pause collecting signatures on petitions but not fully stop like we
do when an election is called. This change allows collection to be
paused and the deadline for petitions open at the time to be extended
by 1 day if the signature collection is paused at midnight.

It also allows for the messaging around any pause to be used
independently of actually pausing collection should it be needed for
other purposes such as making people aware of moderation delays, etc.