Ensure that Rack::ETag doesn't buffer responses
In rack/rack#1416 the Rack::ETag middleware changed it's behaviour
to buffer the response and calculate an ETag digest even though
the Cache-Control: no-cache header is set. This breaks the CSV
streaming responses so to avoid that we use the hack outlined in
rack/rack#1619 and set a Last-Modified header which triggers the
skip_caching? condition in the Rack::ETag middleware.