Fix response and debate threshold methods
In alphagov/e-petitions#691 the signature count update was split into
two updates, one for the count itself and one for the threshold
timestamps. This means that the `at_threshold` methods for response
and debate should now not be one less than the threshold as the count
has been updated and the petition reloaded.