Add print stylesheet
The aim is not to reproduce the site 100% in print but to provide
useful output, e.g. things like forms, buttons, etc. are hidden since
they serve no purpose when printed out.

Not all pages have been tested but the main ones that are likely to
be printed have been.