Don't use Integer() to parse date components
Since Integer() treats leading zeros as meaning this number is octal
then using to parse a component like "09" will break. Switch to using
a regular expression with some whitespace allowed around numbers.

Also workaround the fact that cucumber will convert number strings to
actual numbers so using 'fill in with "09"' doesn't trigger the error
since it parses the number and enters that.