Pause referral or rejection for 24 hours after closing
We allow up to 24 hours for links in emails to be clicked and
signatures validated as a consequence so we can't be 100% sure a
petition has failed to reach the deadline until after this period.

These petitions will only be visible in the all petitions list and
when accessed directly they display a holding message informing the
user that a decision is pending.