Quantize the trending petitions count
Because of the interplay between the CloudFront caching and the
memcached caching the trending petition counts can exhibit weird
behaviour depending on when it's been generated. To prevent this
we quantize the counts into five minute intervals and add the
time to the cache key so that fetches should be consistent.