Return 400 Bad Request for non-integer ids
Most errors reported to Appsignal are ArgumentError exceptions from
id params that fail in Integer(). There's not much to do with them
so we might as well return a 400 Bad Request instead of a 500 error.
This will be ignored by Appsignal and we won't see pointless alerts
in the Slack notification channel.