Pass thresholds in signature count JSON response
The signature count updater resets the thresholds because an update
may take the signature count across either the referral or debate
thresholds. Unfortunately these threshold values are set when the
assets are compiled so reset back to the values in effect when the
deployment occurred. Fix this by passing the thresholds as part of
the JSON response and update where necessary.