Add platform specific versions of nokogiri
As of v1.11.0, nokogiri started shipping native gems but to get this
to work with developing on a Mac (x86 or ARM) we need to add the
plaforms to the lockfile and tweak our package building to include the
linux version which our servers need. This makes our package about 15MB
larger but ours is relatively small any way (45MB after this change).