Increment closed_at directly
Rather than add an extra column, we can increment the closed_at column
directly since that is already populated when a petition is published.

This makes the system more transparent as unlike the UK, the Welsh
petitions website allows customisation of the closing date and having
a hidden value affect that would be confusing for moderators.