Add unreject petitions feature
When petitions are rejected it's meant to be a final state but sometimes
mistakes are made and creators appeal so there are increasing manual
requests to unreject petitions that require developer assistance.

By using the existing 'Restore' functionality we can give the ability
to unreject petitions to moderators and remove developers as a block.