Redirect to the petition page if sponsor token is invalid
Every now and then Facebook spams the site with requests to the new
sponsor url with hundreds of variants of the token. It's unclear why
it's doing this and blocking the ip addresses it uses for this prevents
the link sharing from working so redirect to the relevant petition page
if this happens instead of returning a 404. This prevents the CloudWatch
alarm for 4XX requests being triggered unnecessarily.