Prevent multiple emails when on sponsor threshold
There's a known race condition where two or more sponsors support a
petition at the same time when it's on the threshold for moderation and
multiple copies of the threshold email get sent. This is acceptable as
they're coming from different sponsors. However due to the movement of
signature counting to an external process as part of the aftermath of
the 'Revoke Article 50' petition we inadvertently introduced another
race condition where clicking the verification link again will send
another copy of the threshold email for the same sponsor. We can fix
this by utilising the existing mechanism for preventing the normal
'sponsor supported your petition' email being sent multiple times.