Add removed petitions as an actual state
For various reasons it becomes necessary to remove a petition from
public view after it has been made available, either via collecting
signatures or being rejected for a public reason. To date we have just
been using the rejection feature to change the moderation state to a
hidden reason. However, this only works for petitions that have been
collecting signatures since we need the open_at timestamp to detect
whether the petition has been removed. To make this work for rejected
petitions a developer has to manually touch the open_at timestamp.
Finally, this wasn't an option at all for archived petitions which have
no way to change the moderation of a petition for obvious reasons.

Since this is now occurring more frequently it is now time to add a
'removed' petition state and provide system administrators a UI to
to remove petitions - both for current and archived petitions. When a
petition has been removed it will return a '410 Gone' response in an
attempt to get search engines and other bots to remove it from their
indexes. It's acknowledged that this isn't really going to remove a
petition from the internet due archive.org and other tools that
archive petition data but it should remove it from the results of all
the major search engines.