Allow petitions that have been taken down to be restored
Sometime petitions are mistakenly published and then taken down but
are fine on further checking so allow them to be republished without
the creator having to resubmit the petition again.

Since the restoration process resets the open_at and closed_at
timestamps the feature is limited to sysadmins as it could be abused.