Quote string columns in CSV responses
The Facebook crawler sees the url in a CSV response and attempts to follow
it but includes the signature counts in the line as the comma is not
detected as the end of the url. By quoting the string columns (action, url
and state) it should help badly designed crawlers detect the end of the
url and reduce the spikes of "400 Bad Request" errors we get.