Ensure sequence types are correct
Depending on what version of Rails and PostgreSQL was being used at the time
the sequence type could be either integer or bigint. We keep the sequences
for departments, tags and topics as integer since we're keeping the primary
keys as integer for compatibility with the int_array extension. We also set
the MAXVALUE on archived_petitions to ensure it doesn't run into the sequence
values for the current petitions.