Reset faker unique generator between spec files
The region factory has a requirement for a unique name which is generated
by the county faker and if the specs are run as a whole suite locally then
this can be exhausted. Fix this by resetting the faker unique cache when
running a new spec file as each file should be independent.