Make signature counting more robust
When the RDS service did a security update the inability of delayed_job
to recover from connection errors led to two instances of the counting
job being enqueued as there's a task that runs every 15 minutes to
ensure it's running. When delayed_job was restarted after log rotation
this unlocked the original job and resulted in both jobs being processed
and signatures being double counted for a short period.

This commit tries to prevent this from reoccurring in the future by
using PostgreSQL advisory locks to ensure that two jobs can't run
concurrently in the same way Active Record uses them to prevent
migrations from being run concurrently. Given this restriction we can
put the count job onto it's own queue and if we detect there is more than
one job on the queue then skip processing and don't requeue.

We also change the job to send exceptions to Appsignal as the problem
went unnoticed for a couple of hours - it's not expected that this job
will be continually raising errors so anything occurring should be
addressed as a matter of priority.

This change has been running in production since 2021 on Scottish Petitions
and since 2022 on Welsh Petitions with no adverse effects observed during
the increased demand from the 20mph petition in Wales.

The problem hasn't been observed on the UK site since the change to using
systemd scripts which is probably due to improved handling of errors.