Fix redirect loop with unknown requested content types
Back in 2015 when the site relaunched after the general election there
were a lot of 406 Not Acceptable requests where social media was
interpreting the run-on of a petition url at the end of a sentence with
the first word as the file extension, e.g.

  ... sign petition.parliament.uk/petitions/100001.The ...

To improve the user experience we redirected to the url without the
extension if it wasn't one of our recognised types, e.g. HTML, JSON or
CSV. However in doing so we created a redirect loop where a bot sends
a request with an unregistered content type as it would redirect back
to the url it just requested.

To fix this we only redirect when the url has an extension.